requirements.txt
discord
ffmpeg-python
yt_dlp
beautifulsoup4
python-dotenv
pyshorteners
discord.py[voice]